chore: move process cwd mocking to its own hook for re-use
